requirements.txt
anyio==3.4.0
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.9
dadata==21.10.1
dj-database-url==0.5.0
Django==3.2.9
django-ckeditor==6.2.0
django-debug-toolbar==3.2.2
django-js-asset==1.2.2
django-mptt==0.13.4
django-phonenumber-field==6.0.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.14
django-webp==2.0.0
djangorestframework==3.12.4
gunicorn==20.1.0
h11==0.12.0
httpagentparser==1.9.1
httpcore==0.14.3
httpx==0.21.1
idna==3.3
phonenumberslite==8.12.39
Pillow==9.0.0
psycopg2==2.9.3
pytz==2021.3
requests==2.26.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
sqlparse==0.4.2
urllib3==1.26.7
whitenoise==5.3.0
